+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Added
+ - Add image CAPTCHA support (`complete_image`)
 
 
 ## [0.2.0] - 2020-10-03